MT#56471 add flag to suppress early media
Track audio writes in the mix buffer to set the `active` flag to true
whenever a write occurs, which makes it possible to create the buffer in
an inactivate state and implicitly set it active on demand.

Handle the mix buffer not returning any data in the RTP sending logic
(which is what happens for an inactive buffer) by simply not sending any
packets.

Change-Id: Iaeb0f6deadb3d90020c8c62872735cc94db80504